third_party/blink/web_tests/external/wpt/css/css-inline/baseline-source/baseline-source-last-textarea-001.tentative-expected.txt

This is a testharness.js-based test.
FAIL .target > * 1 assert_equals: 
<span data-offset-y="80"></span>
offsetTop expected 80 but got 29
PASS .target > * 2
PASS .target > * 3
PASS .target > * 4
FAIL .target > * 5 assert_equals: 
<span data-offset-y="80"></span>
offsetTop expected 80 but got 29
PASS .target > * 6
PASS .target > * 7
PASS .target > * 8
PASS .target > * 9
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
Harness: the test ran to completion.
